requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Dance
Flask-Admin
python-dotenv
